requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
boto3==1.37.11
botocore==1.37.11
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dj-database-url==2.3.0
Django==4.2.20
django-bootstrap-v5==1.0.11
django-extensions==4.1
django-storages==1.14.5
et_xmlfile==2.0.0
eval_type_backport==0.2.2
flower==2.0.1
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.2
idna==3.10
jmespath==1.0.1
jsonpath-python==1.0.6
kombu==5.4.2
line_profiler==4.2.0
MarkupSafe==3.0.2
mistralai==1.5.1
mypy-extensions==1.0.0
oauth2client==4.1.3
openpyxl==3.1.5
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyOpenSSL==25.0.0
pyparsing==3.2.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
redis==5.2.1
requests==2.32.3
rsa==4.9
ruff==0.9.7
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
